+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+- Added `logChildElementEvents` config option (defaults to `false`)
+
+### Fixed
+- Fixed deleted element behaviour in multisite
+
 ## 3.0.0 - 2022-10-03
 
 ### Added